Review fixes, added Receiver

+++ contracts/bridge/NativeTreasury.sol
@@ -21,7 +21,7 @@ contract NativeTreasury is INativeTreasury, AccessControlEnumerable  {
         uint256 value_,
         bytes memory data,
         address toSend,
-        uint256 chainIdTo,
+        uint64 chainIdTo,
         address toCall,
         uint256[][] memory valueToSpend,
         bytes[] memory comissionLZ
